requirements.txt
google-adk>=1.0.0
mediapipe==0.10.9
opencv-python>=4.8.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
boto3>=1.29.0
numpy>=1.24.0
pydantic>=2.0.0

